Fix compilation/linking issues
Add rpath to the clang++ call, otherwise the linker does not find the
libraries provided in the builddir, e.g. with

    /usr/bin/ld: warning: libc++.so, needed by /src/v8/v8/out.gn/learning/libv8.so, not found (try using -rpath or -rpath-link)

Disable runtime information with -fno-rtti, otherwise the linker does
not find the typeinfo for the allocator:

    undefined reference to `typeinfo for v8::ArrayBuffer::Allocator'

Background: V8 (and also Node.js) is built without RTTI, but g++ and
clang++ enable them by default. Either V8 has to be build with RTTI
(use_rtti=true) or the own executable without RTTI.