build: use g++ instead of V8's custom libc++
This commit switches to using g++, clang but not using the version
shipped with V8 would also work), as this simplifies things when
compiling and linking. I'm also working on an issue that only manifests
when using g++ so that is also a motivating factor.

The Makefile is a mess right now but I'll clean that up in future
commits. With this commit unit test are working again.